+++ bpf-loader-rs/bpf-loader-lib/src/skeleton/preload/mod.rs
@@ -137,6 +137,7 @@ impl PreLoadBpfSkeleton {
 }
 
 #[cfg(test)]
+#[cfg(not(feature = "no-load-bpf-tests"))]
 mod tests {
     use crate::{
         meta::ComposedObject, skeleton::builder::BpfSkeletonBuilder, tests::get_assets_dir,
