+++ ecli/src/oci/mod.rs
@@ -3,6 +3,7 @@
 //! Copyright (c) 2023, eunomia-bpf
 //! All rights reserved.
 //!
+/// Authencation related stuff
 pub mod auth;
 mod wasm;
 
@@ -18,6 +19,7 @@ use tokio::{
     io::AsyncWriteExt,
 };
 use url::Url;
+/// Ex-export some helper functions from wasm module
 pub use wasm::{parse_img_url, wasm_pull};
 
 use crate::error::{EcliError, EcliResult};
@@ -26,6 +28,7 @@ use wasm::wasm_push;
 
 use self::wasm::{pull::PullArgs, push::PushArgs};
 
+/// A helper function to get the default port for http or https
 pub fn default_schema_port(schema: &str) -> EcliResult<u16> {
     match schema {
         "http" => Ok(80),
@@ -34,6 +37,7 @@ pub fn default_schema_port(schema: &str) -> EcliResult<u16> {
     }
 }
 
+/// Create a HTTP client for the given url
 pub fn get_client(url: &Url) -> EcliResult<Client> {
     Ok(Client::new(ClientConfig {
         protocol: match url.scheme() {
@@ -52,10 +56,12 @@ pub fn get_client(url: &Url) -> EcliResult<Client> {
     }))
 }
 
+/// Push an image to the OCI registry
 pub async fn push(args: PushArgs) -> EcliResult<()> {
     wasm_push(args.file, args.image_url).await
 }
 
+/// Pull an image from the registry
 pub async fn pull(args: PullArgs) -> EcliResult<()> {
     let path = Path::new(args.write_file.as_str());
     let mut file = if !path.exists() {
