+++ ecli/src/oci/wasm/mod.rs
@@ -19,10 +19,13 @@ use self::push::push_wasm_to_registry;
 
 use super::{auth::get_auth_info_by_url, default_schema_port, get_client};
 
+/// Module containing stuff related to pulling image
 pub mod pull;
+/// Module containing stuff related to pushing image
 pub mod push;
 
-// return (..., repo_url_strip_auth_info)
+/// Parse the URL, return things that will be used for pushing / pulling
+/// returns (..., repo_url_strip_auth_info)
 pub fn parse_img_url(url: &str) -> EcliResult<(Client, RegistryAuth, Reference, String)> {
     let img_url = Url::parse(url).map_err(|e| EcliError::ParamErr(e.to_string()))?;
     let auth = match get_auth_info_by_url(&img_url) {
@@ -63,6 +66,7 @@ pub fn parse_img_url(url: &str) -> EcliResult<(Client, RegistryAuth, Reference, 
     ))
 }
 
+/// Push an image
 pub async fn wasm_push(file: String, img_url: String) -> EcliResult<()> {
     // TODO check the file is valid wasm file
     let path = Path::new(file.as_str());
@@ -88,6 +92,7 @@ pub async fn wasm_push(file: String, img_url: String) -> EcliResult<()> {
     Ok(())
 }
 
+/// Pull an image
 pub async fn wasm_pull(img: &str) -> EcliResult<Vec<u8>> {
     let (mut client, auth, reference, repo_url) = parse_img_url(img)?;
     info!("pulling from {}", repo_url);
